+++ examples/mouse_to_tile.rs
@@ -15,10 +15,10 @@ use helpers::camera::movement as camera_movement;
 const MAP_SIDE_LENGTH_X: u32 = 4;
 const MAP_SIDE_LENGTH_Y: u32 = 4;
 
-const TILE_SIZE_SQUARE: TilemapTileSize = TilemapTileSize { x: 50.0, y: 50.0 };
-const TILE_SIZE_ISO: TilemapTileSize = TilemapTileSize { x: 100.0, y: 50.0 };
-const TILE_SIZE_HEX_ROW: TilemapTileSize = TilemapTileSize { x: 50.0, y: 58.0 };
-const TILE_SIZE_HEX_COL: TilemapTileSize = TilemapTileSize { x: 58.0, y: 50.0 };
+const TILE_SIZE_SQUARE: TilemapTileSize = TilemapTileSize { x: 50, y: 50 };
+const TILE_SIZE_ISO: TilemapTileSize = TilemapTileSize { x: 100, y: 50 };
+const TILE_SIZE_HEX_ROW: TilemapTileSize = TilemapTileSize { x: 50, y: 58 };
+const TILE_SIZE_HEX_COL: TilemapTileSize = TilemapTileSize { x: 58, y: 50 };
 const GRID_SIZE_SQUARE: TilemapGridSize = TilemapGridSize { x: 50.0, y: 50.0 };
 const GRID_SIZE_HEX_ROW: TilemapGridSize = TilemapGridSize { x: 50.0, y: 58.0 };
 const GRID_SIZE_HEX_COL: TilemapGridSize = TilemapGridSize { x: 58.0, y: 50.0 };
@@ -81,7 +81,7 @@ fn spawn_tilemap(mut commands: Commands, tile_handle_square: Res<TileHandleSquar
             grid_size,
             size: total_size,
             storage: tile_storage,
-            texture: TilemapTexture(tile_handle_square.clone()),
+            texture: TilemapTexture::Single(tile_handle_square.clone()),
             tile_size,
             map_type: TilemapType::Square {
                 diagonal_neighbors: false,
@@ -200,7 +200,7 @@ fn swap_map_type(
                         diagonal_neighbors: false,
                         coord_system: IsoCoordSystem::Diamond,
                     };
-                    *map_texture = TilemapTexture((*tile_handle_iso).clone());
+                    *map_texture = TilemapTexture::Single((*tile_handle_iso).clone());
                     *tile_size = TILE_SIZE_ISO;
                     *grid_size = GRID_SIZE_ISO;
                 }
@@ -212,7 +212,7 @@ fn swap_map_type(
                         diagonal_neighbors: false,
                         coord_system: IsoCoordSystem::Staggered,
                     };
-                    *map_texture = TilemapTexture((*tile_handle_iso).clone());
+                    *map_texture = TilemapTexture::Single((*tile_handle_iso).clone());
                     *tile_size = TILE_SIZE_ISO;
                     *grid_size = GRID_SIZE_ISO;
                 }
@@ -221,7 +221,7 @@ fn swap_map_type(
                     ..
                 } => {
                     *map_type = TilemapType::Hexagon(HexCoordSystem::Row);
-                    *map_texture = TilemapTexture((*tile_handle_hex_row).clone());
+                    *map_texture = TilemapTexture::Single((*tile_handle_hex_row).clone());
                     *tile_size = TILE_SIZE_HEX_ROW;
                     *grid_size = GRID_SIZE_HEX_ROW;
                 }
@@ -233,7 +233,7 @@ fn swap_map_type(
                 }
                 TilemapType::Hexagon(HexCoordSystem::RowOdd) => {
                     *map_type = TilemapType::Hexagon(HexCoordSystem::Column);
-                    *map_texture = TilemapTexture((*tile_handle_hex_col).clone());
+                    *map_texture = TilemapTexture::Single((*tile_handle_hex_col).clone());
                     *tile_size = TILE_SIZE_HEX_COL;
                     *grid_size = GRID_SIZE_HEX_COL;
                 }
@@ -247,7 +247,7 @@ fn swap_map_type(
                     *map_type = TilemapType::Square {
                         diagonal_neighbors: false,
                     };
-                    *map_texture = TilemapTexture((*tile_handle_square).clone());
+                    *map_texture = TilemapTexture::Single((*tile_handle_square).clone());
                     *tile_size = TILE_SIZE_SQUARE;
                     *grid_size = GRID_SIZE_SQUARE;
                 }
